refactor: move `useFontLoading` code into `FontLoader.vue`
